システム管理者によるアクセス追加

+++ app/controllers/videos_controller.rb
@@ -1,12 +1,17 @@
 class VideosController < ApplicationController
-  before_action :authenticate_user!, only: %i[index new create]
+  before_action :authenticate_user_or_system_admin, only: %i[index]
+  before_action :authenticate_user!, only: %i[new create]
   before_action :set_video, only: %i[show edit update destroy]
   # 投稿者本人またはオーナー→edit, update(追加予定)
-  before_action :ensure_owner, only: %i[destroy]
+  before_action :ensure_owner_or_system_admin, only: %i[destroy]
 
   def index
-    # n+1問題対応.includes([:video_blob])
-    @organization_videos = Video.includes([:video_blob]).current_owner_has(current_user)
+    if user_signed_in?
+      # n+1問題対応.includes([:video_blob])
+      @organization_videos = Video.includes([:video_blob]).current_owner_has(current_user)
+    elsif system_admin_signed_in?
+      @videos = Video.includes([:video_blob]).all
+    end
   end
 
   def new
@@ -34,9 +39,16 @@ class VideosController < ApplicationController
   def update; end
 
   def destroy
-    if @video.owner_has?(current_user) && @video.destroy
-      flash[:danger] = '動画を削除しました'
-      redirect_to videos_path
+    if user_signed_in?
+      if @video.owner_has?(current_user) && @video.destroy
+        flash[:danger] = '動画を削除しました'
+        redirect_to videos_path
+      end
+    elsif system_admin_signed_in?
+      if @video.destroy
+        flash[:danger] = '動画を削除しました'
+        redirect_to videos_path
+      end
     end
   end
 
@@ -47,14 +59,22 @@ class VideosController < ApplicationController
       :popup_after_video)
   end
 
+  def authenticate_user_or_system_admin
+    unless user_signed_in? || system_admin_signed_in? 
+      redirect_to new_user_session_url
+    end
+  end
+
   def set_video
     @video = Video.find(params[:id])
   end
 
-  def ensure_owner
-    if current_user.role != 'owner'
-      flash[:danger] = '権限がありません'
-      redirect_to users_path
+  def ensure_owner_or_system_admin
+    if user_signed_in?
+      if current_user.role != 'owner' 
+        flash[:danger] = '権限がありません'
+        redirect_to users_path
+      end 
     end
   end
 end
